pyproject.toml
[project]
name = "exif-setter"
version = "0.1.0"
description = ""
authors = [
    {name = "charlesl46",email = "charles.lucas4665@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "piexif (>=1.1.3,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "rich (>=14.0.0,<15.0.0)",
]

[tool.poetry.scripts]
exif-setter = "exif_setter.cli:cli"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
